+++ SJYouTubeIDParser Demo/SJYouTubeIDParser Demo/AppDelegate.h
@@ -12,6 +12,21 @@
     
     __weak NSTextField *_videoIDLabel;
     __weak NSTextField *_textField;
+
+    __weak NSTextField *_videoTitleTextField;
+
+    __weak NSProgressIndicator *_circularProgressIndicator;
+
+    __weak NSTextField *_descriptionTextField;
+
+
+    __weak NSTextField *_timeAgoTextField;
+    __weak NSImageView *_thumbnailImageView;
+
+    __weak NSTextField *_uploaderTextField;
+    __weak NSLevelIndicator *_ratingIndicator;
+
+    __weak NSTextField *_viewCountTextField;
 }
 
 @property (assign) IBOutlet NSWindow *window;
@@ -19,4 +34,12 @@
 @property (weak) IBOutlet NSTextField *textField;
 - (IBAction)extractAction:(id)sender;
 @property (weak) IBOutlet NSTextField *videoIDLabel;
+@property (weak) IBOutlet NSTextField *videoTitleTextField;
+@property (weak) IBOutlet NSTextField *descriptionTextField;
+@property (weak) IBOutlet NSImageView *thumbnailImageView;
+@property (weak) IBOutlet NSTextField *viewCountTextField;
+@property (weak) IBOutlet NSLevelIndicator *ratingIndicator;
+@property (weak) IBOutlet NSProgressIndicator *circularProgressIndicator;
+@property (weak) IBOutlet NSTextField *timeAgoTextField;
+@property (weak) IBOutlet NSTextField *uploaderTextField;
 @end