+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+#### Version 1.1.0
+
+* Added block methods
+* Updated Demo Project
+* When calling "getVideoInfoFromURL" you can now get YouTube video's
+video ID, thumbnail URL, video description, view count,rating, upload date, and the uploader
+* Added + defaultParser instance
+* Fixed all known issues
+* Now requires AFNetworking
+
+
+#### Version 1.1.0 bc1
+
+* Added blocks
+* Updated Demo Project
+* When calling "getVideoInfoFromURL" you can now get YouTube video's
+video ID, thumbnail URL, video description, view count, and the rating
+* Added + defaultParser instance
+
+**Known issues:**
+* Parser will crash if attemping to request video info with no internet
+connection.
 
 #### Version 1.0.0